// apps/ui/src/lib/api-client/models/AppStoreApp.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart Backend
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.6-alpha.202604171740.60503a70
 *
 * NOTE: This class is auto generated by generate-ts-fetch-client.py
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';

/**
 * @export
 * @interface AppStoreApp
 */
export interface AppStoreApp {
    /**
     * @type {string}
     * @memberof AppStoreApp
     */
    id: string;
    /**
     * @type {string}
     * @memberof AppStoreApp
     */
    launchUrl: string;
    /**
     * @type {string}
     * @memberof AppStoreApp
     */
    clientId: string;
    /**
     * @type {string}
     * @memberof AppStoreApp
     */
    clientName: string;
    /**
     * @type {string}
     * @memberof AppStoreApp
     */
    description: string;
    /**
     * @type {string}
     * @memberof AppStoreApp
     */
    category: string;
    /**
     * @type {string}
     * @memberof AppStoreApp
     */
    icon: string;
    /**
     * @type {boolean}
     * @memberof AppStoreApp
     */
    hidden: boolean;
    /**
     * @type {AppStoreAppSourceEnum}
     * @memberof AppStoreApp
     */
    source: AppStoreAppSourceEnum;
}

/**
 * @export
 */
export const AppStoreAppSourceEnum = {
    Filesystem: 'filesystem',
    Registered: 'registered',
} as const;
export type AppStoreAppSourceEnum = typeof AppStoreAppSourceEnum[keyof typeof AppStoreAppSourceEnum];

/**
 * Check if a given object implements the AppStoreApp interface.
 */
export function instanceOfAppStoreApp(value: object): value is AppStoreApp {
    if (!('clientId' in value) || value['clientId'] === undefined) return false;
    if (!('category' in value) || value['category'] === undefined) return false;
    if (!('icon' in value) || value['icon'] === undefined) return false;
    if (!('hidden' in value) || value['hidden'] === undefined) return false;
    if (!('clientName' in value) || value['clientName'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('source' in value) || value['source'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('launchUrl' in value) || value['launchUrl'] === undefined) return false;
    return true;
}

export function AppStoreAppFromJSON(json: any): AppStoreApp {
    return AppStoreAppFromJSONTyped(json, false);
}

export function AppStoreAppFromJSONTyped(json: any, ignoreDiscriminator: boolean): AppStoreApp {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'launchUrl': json['launch_url'],
        'clientId': json['client_id'],
        'clientName': json['client_name'],
        'description': json['description'],
        'category': json['category'],
        'icon': json['icon'],
        'hidden': json['hidden'],
        'source': json['source'],
    };
}

export function AppStoreAppToJSON(json: any): AppStoreApp {
    return AppStoreAppToJSONTyped(json, false);
}

export function AppStoreAppToJSONTyped(value?: AppStoreApp | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'launch_url': value['launchUrl'],
        'client_id': value['clientId'],
        'client_name': value['clientName'],
        'description': value['description'],
        'category': value['category'],
        'icon': value['icon'],
        'hidden': value['hidden'],
        'source': value['source'],
    };
}