// apps/ui/src/lib/api-client/models/PublishedApp.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Proxy Smart Backend
 * SMART on FHIR Proxy + Healthcare Administration API using Keycloak and Elysia
 *
 * The version of the OpenAPI document: 0.0.6-alpha.202604171740.60503a70
 *
 * NOTE: This class is auto generated by generate-ts-fetch-client.py
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';

/**
 * @export
 * @interface PublishedApp
 */
export interface PublishedApp {
    /**
     * @type {string}
     * @memberof PublishedApp
     */
    clientId: string;
    /**
     * @type {string}
     * @memberof PublishedApp
     */
    name: string;
    /**
     * @type {string}
     * @memberof PublishedApp
     */
    description: string;
    /**
     * @type {string}
     * @memberof PublishedApp
     */
    launchUrl: string;
    /**
     * @type {string}
     * @memberof PublishedApp
     */
    category: string;
    /**
     * @type {string}
     * @memberof PublishedApp
     */
    logoUri?: string;
}

/**
 * Check if a given object implements the PublishedApp interface.
 */
export function instanceOfPublishedApp(value: object): value is PublishedApp {
    if (!('category' in value) || value['category'] === undefined) return false;
    if (!('clientId' in value) || value['clientId'] === undefined) return false;
    if (!('description' in value) || value['description'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('launchUrl' in value) || value['launchUrl'] === undefined) return false;
    return true;
}

export function PublishedAppFromJSON(json: any): PublishedApp {
    return PublishedAppFromJSONTyped(json, false);
}

export function PublishedAppFromJSONTyped(json: any, ignoreDiscriminator: boolean): PublishedApp {
    if (json == null) {
        return json;
    }
    return {
        
        'clientId': json['clientId'],
        'name': json['name'],
        'description': json['description'],
        'launchUrl': json['launchUrl'],
        'category': json['category'],
        'logoUri': json['logoUri'] == null ? undefined : json['logoUri'],
    };
}

export function PublishedAppToJSON(json: any): PublishedApp {
    return PublishedAppToJSONTyped(json, false);
}

export function PublishedAppToJSONTyped(value?: PublishedApp | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'clientId': value['clientId'],
        'name': value['name'],
        'description': value['description'],
        'launchUrl': value['launchUrl'],
        'category': value['category'],
        'logoUri': value['logoUri'],
    };
}